Try to fix whitespace sensitivity in header links
Inside a JSX expression, the prettier plugin for astro was adding
newlines around the content, which resulted in visible whitespace (i.e.,
extra space for underlines).

The solution in HeaderLink was to actually drop the use of a conditional
in JSX and instead always add a area-current attribute, but set it
"false" when necessary.